refactor(components/table): move filter to xm-table
